+++ packages/email-notification-dispatcher/src/handlers/eservices/handleEserviceEvent.ts
@@ -5,7 +5,7 @@ import {
 import { match, P } from "ts-pattern";
 import { config } from "../../config/config.js";
 import { HandlerParams } from "../../models/handlerParams.js";
-import { handleEserviceDescriptorPublished } from "./handleEServiceDescriptorPublished.js";
+import { handleEserviceDescriptorPublished } from "./handleEserviceDescriptorPublished.js";
 
 const interopFeBaseUrl = config.interopFeBaseUrl;
