+++ packages/email-notification-dispatcher/test/handleEServiceDescriptorPublished.test.ts
@@ -21,7 +21,7 @@ import {
   descriptorPublishedNotFound,
   tenantNotFound,
 } from "../src/models/errors.js";
-import { handleEserviceDescriptorPublished } from "../src/handlers/eservices/handleEServiceDescriptorPublished.js";
+import { handleEserviceDescriptorPublished } from "../src/handlers/eservices/handleEserviceDescriptorPublished.js";
 import {
   addOneAgreement,
   addOneEService,
